4/2/2025
casa

+++ PSP/Intellij/UD1/ProcessBuilder/src/main/java/org/example/REPASO/encryptProgram/encryptProgram.java
@@ -30,7 +30,7 @@ public class encryptProgram {
             char c= caracteres2[index];
 
             if (caracteres2[index] >= (caracteres2.length/2)){
-                int caracterNum2= (int)c;
+                int caracterNum2= c;
                 caracterNum2-=1;
                 sbFinal.append((char)caracterNum2);
             }else {
